+++ backend/src/database/seeds/seed.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { ConfigModule } from '@nestjs/config';
+import { User } from 'user/user.entity';
+import { SeedService } from './seed.service';
+
+@Module({
+  imports: [ConfigModule, TypeOrmModule.forFeature([User])],
+  providers: [SeedService],
+  exports: [SeedService],
+})
+export class SeedModule {}